Base datos

+++ backend-lugares-turisticos/src/controllers/usuario.controller.ts
@@ -6,6 +6,7 @@ import {FileInterceptor} from "@nestjs/platform-express";
 import {CommonSchema} from "../schemas/common.schema";
 import {EntityPipe} from "../pipes/entity.pipe";
 import {RolesGuard} from "../guards/auth.guard";
+import {ErrorIngresoDatosException} from "../exceptions/error-ingreso-datos.exception";
 
 
 @Controller('usuario')
@@ -15,8 +16,17 @@ export class UsuarioController {
     }
 
     @Post()
-    create(@Body(new EntityPipe(CommonSchema.USUARIO_SCHEMA)) crearUsuario) {
-        return this._usuarioService.insert(crearUsuario);
+    create(@Body(new EntityPipe(CommonSchema.USUARIO_SCHEMA)) crearUsuario,@Res() response) {
+        return this._usuarioService.insert(crearUsuario)
+            .then(()=> response.status(200).json(
+                {
+                    data: crearUsuario
+                }))
+            .catch(err=> {
+                if(err){
+                    throw new  ErrorIngresoDatosException(err.message,err.detail);
+                }
+        })
     }
 
     @UseGuards(RolesGuard)
@@ -29,19 +39,41 @@ export class UsuarioController {
     async findOneNick(@Param('correo') correo, @Res() response) {
         return response.send(await this._usuarioService.selectPorCorreo(correo));
     }
+
+
+    @UseGuards(RolesGuard)
     @Get('id/:id')
     async findOne(@Param('id') id, @Res() response) {
         return response.send(await this._usuarioService.selectById(id));
     }
+
+
+    @UseGuards(RolesGuard)
     @Put(':id')
     async update(@Param('id') id, @Body() nuevo) {
-        return await this._usuarioService.update(id, nuevo);
+        var idEncontrado = await this._usuarioService.selectById(id);
+        if(idEncontrado==undefined){
+            throw new  ErrorIngresoDatosException("Usuario no encontrado!!","El id del usuario aun no ha sido registrado");
+
+        }else{
+            return await this._usuarioService.update(id, nuevo);
+
+        }
     }
 
+    @UseGuards(RolesGuard)
     @Delete(':id')
     async remove(@Param('id') id) {
-        return await this._usuarioService.delete(id);
+        var idEncontrado = await this._usuarioService.selectById(id);
+        if(idEncontrado==undefined){
+            throw new  ErrorIngresoDatosException("Usuario no encontrado!!","El id del usuario aun no ha sido registrado");
+
+        }else{
+            return await this._usuarioService.delete(id);
+
+        }
     }
+
 
     @Post('upload-image')
     @UseInterceptors(FileInterceptor('image', {
@@ -90,8 +122,15 @@ export class UsuarioController {
             };
 
         } else {
-            throw new HttpException('¡Formato de imagen no soportado!', HttpStatus.BAD_REQUEST);
+            throw new HttpException('¡Elija una imagen!', HttpStatus.BAD_REQUEST);
         }
+    }
+
+
+    @UseGuards(RolesGuard)
+    @Get('buscarFoto/:archivo')
+    async findFile(@Res() response, @Param('archivo') archivo){
+        return response.sendFile(`users/${archivo}`,{root: 'public'});
     }
 
 }
